at this point we have a square being drawn on the demo page
