+++ docs/index.html
@@ -6,7 +6,7 @@
   <meta name="viewport" content="width=device-width, initial-scale=1.0">
   <meta http-equiv="X-UA-Compatible" content="ie=edge">
   <meta name="description" content="CHANGE THIS">
-  <title>React component starter</title>
+  <title>iaReact component starter</title>
 </head>
 
 <body>
