+++ src/docs/index.jsx
@@ -1,13 +1,13 @@
 import React from "react";
 import { render } from "react-dom";
-import MyComponent from "../../lib";
+import WingTron from "../../lib/index.js";
 import "./styles.css";
 
 function Demo() {
   return (
     <div>
       <h1>Demo with examples of the component</h1>
-      <MyComponent color="brown">Wow what a button</MyComponent>
+      <WingTron />
     </div>
   );
 }
