--- src/unconverted-code/wingTron.html
@@ -1,21 +1,0 @@
-<!doctype html>
-<html lang="en">
-<head>
-	<meta charset="UTF-8">
-	<title>Tron</title>
-	<meta name="viewport" content="width=device-width, initial-scale=1">
-	<script data-main="../src/main.js" src="../src/require.js"></script>
-<head>
-<body>
-<div class="container">
-	<div class="row">
-		<div class="col-md-2 scoreboard">
-			<h1>Scores</h1>
-			<ol></ol>
-			<button id="gameStartButton" type="button" class="btn btn-default" aria-label="Left Align">Start Game</button>
-		</div>
-		<div class="col-md-10 canvasBlock"></div>
-	</div>
-</div>
-</body>
-</html>
